ref: allocate validator to individual file
