+++ Common/__init__.py
@@ -1,1 +1,2 @@
-
+from .utils import error_message
+from .utils import exit_message
